Change IDN to wave mode, and more (#29)
* Started switch to IDN wave mode, but buggy

* Improve simulated timing

* Fix possible wave packet with too low point count.

* Misc polish

* Improved but still not good enough jitter. Also fix duplicate servers based on unit-id

* Checkpoint before trying ring buffer

* Major refactor, glitchy still

* Small tweaks. Glitchy, probably race condition

* Finally working well

* Polish

* Added more functions for differentiating between USB and IDN devices

* Protect against too small frames

* polish

* Fix possible crash when closing device, and unix timing

* Revert accidental wrong mutex

* Polish

* Sort device list and fix another possible crash when closing IDN device

* More tolerant timing if buffer underrun a little bit

* Don't count late waits when no frame is playing

* Fix possible stuck state after buffer underrun

* Final touches on documentation/comments

* Binaries

+++ sdk/cpp/idn/plt-posix.c
@@ -38,4 +38,4 @@
 
 int plt_monoValid = 0;
 struct timespec plt_monoRef = { 0 };
-uint32_t plt_monoTimeUS = 0;
+uint64_t plt_monoTimeUS = 0;
